+++ ledger-service/http-json/BUILD.bazel
@@ -41,6 +41,7 @@ da_scala_library(
         "//ledger/ledger-api-common",
         "//libs-scala/auth-utils",
         "//libs-scala/ports",
+        "//libs-scala/scala-utils",
         "@maven//:com_chuusai_shapeless_2_12",
         "@maven//:com_github_scopt_scopt_2_12",
         "@maven//:com_lihaoyi_sourcecode_2_12",
@@ -84,6 +85,7 @@ da_scala_binary(
         "//ledger/ledger-api-common",
         "//libs-scala/auth-utils",
         "//libs-scala/ports",
+        "//libs-scala/scala-utils",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_chuusai_shapeless_2_12",
         "@maven//:com_github_scopt_scopt_2_12",
@@ -129,6 +131,7 @@ da_scala_test(
         "//ledger-service/db-backend",
         "//ledger-service/lf-value-json",
         "//ledger/ledger-api-common",
+        "//libs-scala/scala-utils",
         "//libs-scala/scalatest-utils",
         "@maven//:com_chuusai_shapeless_2_12",
         "@maven//:com_lihaoyi_sourcecode_2_12",
@@ -184,6 +187,7 @@ da_scala_test(
         "//libs-scala/auth-utils",
         "//libs-scala/ports",
         "//libs-scala/postgresql-testing",
+        "//libs-scala/scala-utils",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_chuusai_shapeless_2_12",
         "@maven//:com_github_ghik_silencer_lib_2_12_11",
