+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/WebSocketService.scala
@@ -292,7 +292,7 @@ class WebSocketService(
     extends LazyLogging {
 
   import WebSocketService._
-  import Statement.discard
+  import com.daml.scala.util.Statement.discard
   import util.ErrorOps._
   import com.daml.http.json.JsonProtocol._
