+++ libs-scala/scala-utils/src/main/scala/com/daml/scala/util/Statement.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.http
+package com.daml.scala.util
 
 object Statement {
   @specialized def discard[A](evaluateForSideEffectOnly: A): Unit = {
